+++ src/plugin-api/main/uk/co/reecedunn/intellij/plugin/intellij/execution/ui/results/QueryTextConsoleView.kt
@@ -27,6 +27,9 @@ import uk.co.reecedunn.intellij.plugin.core.ui.Borders
 import uk.co.reecedunn.intellij.plugin.intellij.execution.process.QueryProcessHandlerBase
 import uk.co.reecedunn.intellij.plugin.intellij.execution.process.QueryResultListener
 import uk.co.reecedunn.intellij.plugin.intellij.execution.process.QueryResultTime
+import uk.co.reecedunn.intellij.plugin.processor.database.DatabaseModule
+import uk.co.reecedunn.intellij.plugin.processor.database.resolve
+import uk.co.reecedunn.intellij.plugin.processor.debug.createNavigatable
 import uk.co.reecedunn.intellij.plugin.processor.query.QueryError
 import uk.co.reecedunn.intellij.plugin.processor.query.QueryResult
 import uk.co.reecedunn.intellij.plugin.xpath.model.XsDurationValue
@@ -100,10 +103,18 @@ class QueryTextConsoleView(project: Project) : TextConsoleView(project), QueryRe
                 }
             }
             e.frames.forEach {
-                print(
-                    "    at ${it.module ?: ""}:${it.lineNumber}:${it.columnNumber}\n",
-                    ConsoleViewContentType.ERROR_OUTPUT
-                )
+                print("    at ", ConsoleViewContentType.ERROR_OUTPUT)
+                if (it.module != null) {
+                    val resolved = when(it.module) {
+                        is DatabaseModule -> it.module.resolve(project).firstOrNull()
+                        else -> it.module
+                    }
+                    if (resolved == null)
+                        print(it.module.path, ConsoleViewContentType.ERROR_OUTPUT)
+                    else
+                        printHyperlink(it.module.path) { project -> it.createNavigatable(project)?.navigate(true) }
+                }
+                print(":${it.lineNumber}:${it.columnNumber}\n", ConsoleViewContentType.ERROR_OUTPUT)
             }
         } else {
             val writer = StringWriter()